requirements.txt
Django==4.2.3
django-cors-headers==4.3.1
django-filter==23.5
django-phonenumber-field[phonenumberslite]==7.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
factory-boy==3.3.0
Pillow==10.2.0
python-dotenv==1.0.0
setuptools==69.0.3
